Add files via upload
Exploratory Data Analysis is crucial in Data Analysis. It involves loading the dataset, examining the first few rows of the dataset using the head() function, the structure and data types of the dataset using the info(), handling missing values and duplicates and analyzing the relationship between variables and identifying patterns in the dataset.